REFACTOR(#49): restructure backend
-separating graphql types into several files
-splitting resolvers into several files
-graphql link changed to http://localhost:3001/graphql
-removing unnecessary console.log
- adding packages fs, path, @graphql-tools/schema
