+++ graphql/index.js
@@ -1,0 +1,11 @@
+import { ApolloServer } from "apollo-server-express";
+import schema from "./schema.js";
+
+const apolloServer = new ApolloServer({
+    schema,
+    context: ({ req }) => {
+        return req;
+    },
+});
+
+export default apolloServer;
