+++ graphql/resolvers/index.js
@@ -1,0 +1,10 @@
+import { userMutations } from "./user/index.js";
+
+const resolvers = {
+    Query: {},
+    Mutation: {
+        ...userMutations,
+    },
+};
+
+export default resolvers;
