+++ graphql/resolvers/user/index.js
@@ -1,0 +1,1 @@
+export { default as userMutations } from "./mutations.js";
