--- graphql/schema.graphql
@@ -1,16 +1,0 @@
-type Query{
-    query: String!
-}
-
-type Mutation{
-    login(username: String!, password: String!): Token!
-    register(username: String!, password: String!): Token!
-}
-
-type Token{
-    accessToken: String!,
-    refreshToken: String!,
-
-    user_id: String!,
-    username: String!
-}
