+++ graphql/typedefs/mutation.graphql
@@ -1,0 +1,4 @@
+type Mutation {
+    login(username: String!, password: String!): Token!
+    register(username: String!, password: String!): Token!
+}
