+++ graphql/typedefs/query.graphql
@@ -1,0 +1,3 @@
+type Query {
+    tmp_root_query: String!
+}
