+++ package.json
@@ -21,13 +21,16 @@
     },
     "homepage": "https://github.com/Darkosz1012/amazonka#readme",
     "dependencies": {
+        "@graphql-tools/schema": "^7.1.3",
         "apollo-server": "^2.23.0",
         "crypto": "^1.0.1",
         "dotenv": "^8.2.0",
         "express": "^4.17.1",
-        "mongoose": "^5.12.5",
+        "fs": "0.0.1-security",
         "graphql": "^15.5.0",
-        "jsonwebtoken": "^8.5.1"
+        "jsonwebtoken": "^8.5.1",
+        "mongoose": "^5.12.5",
+        "path": "^0.12.7"
     },
     "devDependencies": {
         "@commitlint/cli": "^12.1.1",
